app/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
redis==5.2.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
requests==2.32.3
celery==5.4.0
